Factor out SuggestedLanguage into a separate component
