+++ src/view/com/composer/Composer.tsx
@@ -45,6 +45,7 @@ import {Gallery} from './photos/Gallery'
 import {MAX_GRAPHEME_LENGTH} from 'lib/constants'
 import {LabelsBtn} from './labels/LabelsBtn'
 import {SelectLangBtn} from './select-language/SelectLangBtn'
+import {SuggestedLanguage} from './select-language/SuggestedLanguage'
 import {insertMentionAt} from 'lib/strings/mention-manip'
 import {Trans, msg} from '@lingui/macro'
 import {useLingui} from '@lingui/react'
@@ -62,8 +63,6 @@ import {emitPostCreated} from '#/state/events'
 import {ThreadgateSetting} from '#/state/queries/threadgate'
 import {logger} from '#/logger'
 import {ComposerReplyTo} from 'view/com/composer/ComposerReplyTo'
-import lande from 'lande'
-import {code3ToCode2Strict, codeToLanguageName} from '#/locale/helpers'
 
 type Props = ComposerOpts
 export const ComposePost = observer(function ComposePost({
@@ -182,31 +181,6 @@ export const ComposePost = observer(function ComposePost({
       return () => window.removeEventListener('keydown', onEscape)
     }
   }, [onEscape])
-
-  const [suggestedLanguage, setSuggestedLanguage] = useState<string>()
-  useEffect(() => {
-    const text = richtext.text.trim()
-
-    // Don't run the language model on small posts, the results are likely
-    // to be inaccurate anyway.
-    if (text.length < 10) {
-      setSuggestedLanguage(undefined)
-      return
-    }
-
-    const idle = requestIdleCallback(() => {
-      // Only select languages that are
-      const result = lande(text).filter(
-        ([lang, value]) => value >= 0.85 && code3ToCode2Strict(lang),
-      )
-
-      setSuggestedLanguage(
-        result.length > 0 ? code3ToCode2Strict(result[0][0]) : undefined,
-      )
-    })
-
-    return () => cancelIdleCallback(idle)
-  }, [richtext])
 
   const onPressAddLinkCard = useCallback(
     (uri: string) => {
@@ -481,39 +455,7 @@ export const ComposePost = observer(function ComposePost({
               ))}
           </View>
         ) : null}
-        {suggestedLanguage &&
-        !toPostLanguages(langPrefs.postLanguage).includes(suggestedLanguage) ? (
-          <View style={[pal.borderDark, styles.infoBar]}>
-            <Text style={[pal.text, s.flex1]}>
-              <Trans>
-                You seem to be writing in{' '}
-                <Text type="sm-bold" style={pal.text}>
-                  {codeToLanguageName(suggestedLanguage)}
-                </Text>
-                , but your post language is currently set to{' '}
-                <Text type="sm-bold" style={pal.text}>
-                  {toPostLanguages(langPrefs.postLanguage)
-                    .map(lang => codeToLanguageName(lang))
-                    .join(', ')}
-                </Text>
-              </Trans>
-            </Text>
-
-            <TouchableOpacity
-              onPress={() => setLangPrefs.setPostLanguage(suggestedLanguage)}
-              accessibilityRole="button"
-              accessibilityLabel={_(
-                msg`Switch to ${codeToLanguageName(suggestedLanguage)}`,
-              )}
-              accessibilityHint={_(
-                msg`Switch to ${codeToLanguageName(suggestedLanguage)}`,
-              )}>
-              <Text style={pal.link}>
-                <Trans>Switch</Trans>
-              </Text>
-            </TouchableOpacity>
-          </View>
-        ) : null}
+        <SuggestedLanguage text={richtext.text} />
         <View style={[pal.border, styles.bottomBar]}>
           {canSelectImages ? (
             <>
@@ -624,16 +566,5 @@ const styles = StyleSheet.create({
     paddingRight: 20,
     alignItems: 'center',
     borderTopWidth: 1,
-  },
-
-  infoBar: {
-    flexDirection: 'row',
-    gap: 16,
-    borderWidth: 1,
-    borderRadius: 6,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    marginHorizontal: 10,
-    marginBottom: 10,
   },
 })
